requirements.txt
config
pyside6
pyqt6
numpy
matplotlib
pyserial
pyusb